[#2] Initial work to create donation and donor
A big commit that allows donations to be created alongside donors. I am keeping
the database interaction in one repo at the moment, but the donor code can be
moved to its own repo if needed.

One integration spec has been written that tests the submission of donation to
the API.

Initially it has been decided to use the JSON-API format.